Support docker on agent hosts

+++ docker/entrypoint.sh
@@ -15,7 +15,8 @@ if [ -n "${INIT}" ]; then
   ${INIT}
 fi
 # Check if we have a package, and if so, bootstrap it
-if [ -f /package.uvn-agent ]; then
+# if the ${UVN_DIR} is empty
+if [ -f /package.uvn-agent -a -z "$(find ${UVN_DIR} -mindepth 1 -maxdepth 1)" ]; then
   chmod 600 /package.uvn-agent &&
     uno install ${VERBOSE} /package.uvn-agent -r ${UVN_DIR}
 fi
@@ -23,22 +24,27 @@ fi
 case "$1" in
 # Start an agent for a cell or the registry
 agent)
+  shift
   exec uno agent -r ${UVN_DIR} $@
   ;;
 # Regenerate deployment
 redeploy)
+  shift
   exec uno redeploy -r ${UVN_DIR} $@
   ;;
 # Start static network services
 up)
+  shift
   exec uno service up -r ${UVN_DIR} $@
   ;;
 # Stop static network services
 down)
+  shift
   exec uno service down -r ${UVN_DIR} $@
   ;;
 # Push configuration to cells
 sync)
+  shift
   exec uno sync -r ${UVN_DIR} $@
   ;;
 # Validate arguments and skip to bottom for chown
